Refactoring segue logic and renaming functions/variables
