+++ practise-project/Home/HomeViewController.swift
@@ -15,5 +15,4 @@ class HomeViewController: UIViewController {
         // Do any additional setup after loading the view, typically from a nib.
     }
 
-    
 }
